exercises/ex5/requirements.txt
streamlit==1.27.2
googlesearch-python==1.2.3
beautifulsoup4==4.12.2
PyYAML==6.0.1